# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zendesk API (baton-zendesk connector surface)",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://{subdomain}.zendesk.com/api/v2",
      "description": "Zendesk instance, built from the `subdomain` config field. A `base-url` config value takes precedence over this and replaces the whole URL.",
      "variables": {
        "subdomain": {
          "default": "example"
        }
      }
    }
  ],
  "security": [
    {
      "apiTokenBasic": []
    },
    {
      "oauthClientCredentials": [
        "read",
        "write"
      ]
    }
  ],
...
  "paths": {
...
    "/users/{user_id}.json": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "users"
        ],
        "summary": "Get a user",
        "externalDocs": {
          "url": "https://developer.zendesk.com/api-reference/ticketing/users/users/"
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "User.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "tags": [
          "users"
        ],
        "summary": "Update a user",
        "description": "The connector sends an arbitrary field map rather than a typed user, so the request body is open — whichever attributes the caller wants changed, wrapped in a `user` object.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Passed through as map[string]any. In practice the caller supplies a `user` object carrying the attributes to change.",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "tags": [
          "users"
        ],
        "summary": "Soft-delete a user",
        "description": "The connector treats a 200 reported through the SDK's error type as success, because go-zendesk raises an error for some 200 responses on this endpoint.",
        "externalDocs": {
          "url": "https://developer.zendesk.com/api-reference/ticketing/users/users/#delete-user"
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "User soft-deleted."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiTokenBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "Zendesk API token auth. The username is the configured email with `/token` appended; the password is the API token."
      },
      "oauthClientCredentials": {
        "type": "oauth2",
        "description": "Zendesk OAuth client credentials grant, selected with `auth-method=oauth-client-credentials`. The connector POSTs a JSON body (not a form) to `https://{subdomain}.zendesk.com/oauth/tokens` (or the origin of `base-url`), caches the access token in memory, and sends it as `Authorization: Bearer`. No refresh token is issued; on expiry or a 401 `invalid_token` the connector requests a new token and retries the request once.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://example.zendesk.com/oauth/tokens",
            "scopes": {
              "read": "Read access to all resources the connector syncs.",
              "write": "Write access, needed for provisioning and ticketing."
            }
          }
        }
      }
    },
    "parameters": {
      "UserId": {
        "name": "user_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "description": "The connector decodes go-zendesk's full User struct. Fields it actively reads are id, name, email and active; the rest are decoded and available.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "description": "Drives the resource's enabled/disabled status."
          },
          "alias": {
            "type": "string"
          },
          "chat_only": {
            "type": "boolean"
          },
          "custom_role_id": {
            "type": "integer",
            "format": "int64"
          },
          "default_group_id": {
            "type": "integer",
            "format": "int64"
          },
          "details": {
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "iana_time_zone": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "locale_id": {
            "type": "integer",
            "format": "int64"
          },
          "moderator": {
            "type": "boolean"
          },
          "notes": {
            "type": "string"
          },
          "only_private_comments": {
            "type": "boolean"
          },
          "organization_id": {
            "type": "integer",
            "format": "int64"
          },
          "phone": {
            "type": "string"
          },
          "photo": {
            "type": "object",
            "additionalProperties": true
          },
          "remote_photo_url": {
            "type": "string"
          },
          "restricted_agent": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "enum": [
              "end-user",
              "agent",
              "admin"
            ]
          },
          "role_type": {
            "type": "integer",
            "format": "int64"
          },
          "shared": {
            "type": "boolean"
          },
          "shared_agent": {
            "type": "boolean"
          },
          "shared_phone_number": {
            "type": "boolean"
          },
          "signature": {
            "type": "string"
          },
          "suspended": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ticket_restriction": {
            "type": "string"
          },
          "time_zone": {
            "type": "string"
          },
          "two_factor_auth_enabled": {
            "type": "boolean"
          },
          "user_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "verified": {
            "type": "boolean"
          },
          "report_csv": {
            "type": "boolean"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}